tests/requirements.txt
pytest
pytest-cov
pyfakefs
freezegun
pytest-flask
playwright
pytest-playwright
requests
pytest-httpserver
docker
